requirements.txt
# Helios Trading System V3.0 - Dependencies
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
influxdb-client==1.39.0
alembic==1.13.1
sqlalchemy==2.0.25

# ML/AI
torch==2.1.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.4.0
ta==0.11.0
arch==6.3.0

# LLM
anthropic==0.18.1
openai==1.12.0

# HTTP Client
aiohttp==3.9.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
pytz==2024.1
python-dateutil==2.8.2

# Monitoring
prometheus-client==0.19.0

# Security
cryptography==42.0.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0
